INS-291: move logicrunner's build helper in testutil
